+++ tala/nl/gf/naming.py
@@ -8,3 +8,7 @@ def natural_language_gf_filename(ddd_name, language_code):
 
 def semantic_gf_filename(ddd_name):
     return "%s_sem.gf" % ddd_name
+
+
+def probabilities_filename(ddd_name):
+    return "%s.probs" % ddd_name